requirements.txt
aiohttp==0.22.5
aiopg==0.10.0
alembic==0.8.7
bcrypt==3.1.0
cffi==1.7.0
chardet==2.3.0
Mako==1.0.4
MarkupSafe==0.23
multidict==1.2.2
psycopg2==2.6.2
pycparser==2.14
PyJWT==1.4.2
python-editor==1.0.1
schematics==1.1.1
six==1.10.0
SQLAlchemy==1.0.15
